Fix: separate ApiId and ApiHandler
