+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/web_api/handler/PlayerInfo.kt
@@ -20,12 +20,11 @@ import net.minecraft.server.network.ServerPlayerEntity
 import online.ruin_of_future.informative_mc_core.ImcCore
 import online.ruin_of_future.informative_mc_core.data.ModDataManager
 import online.ruin_of_future.informative_mc_core.web_api.ApiID
+import online.ruin_of_future.informative_mc_core.web_api.ids.PlayerInfoApiId
 import online.ruin_of_future.informative_mc_core.web_api.response.PlayerInfoResponse
 import online.ruin_of_future.informative_mc_core.web_api.response.PlayerInfoResponseBody
 import online.ruin_of_future.informative_mc_core.web_api.response.SinglePlayerInfo
 import java.io.OutputStream
-
-val PlayerInfoApiId = ApiID("mc-info", "player-info")
 
 class PlayerInfoHandler(
     private val modDataManager: ModDataManager,
